autorescalado android y cambio de resolucion al vuelo

+++ Practica_1/Interfaces/src/main/java/gdv/ucm/interfaces/StateManager.java
@@ -2,7 +2,7 @@ package gdv.ucm.interfaces;
 
 public interface StateManager {
 
-    void update(long deltaTime);//delta time aqui
+    void update(float deltaTime);//delta time aqui
     void render();
 
 }
